+++ addons/hr_recruitment/__manifest__.py
@@ -24,7 +24,13 @@
         'data/mail_template_data.xml',
         'data/mail_templates.xml',
         'data/hr_recruitment_data.xml',
-        'views/hr_recruitment_views.xml',
+        'views/hr_recruitment_degree_views.xml',
+        'views/hr_recruitment_source_views.xml',
+        'views/hr_recruitment_stage_views.xml',
+        'views/ir_attachment_views.xml',
+        'views/hr_applicant_category_views.xml',
+        'views/hr_applicant_refuse_reason_views.xml',
+        'views/hr_applicant_views.xml',
         'views/res_config_settings_views.xml',
         'views/hr_department_views.xml',
         'views/hr_job_views.xml',
