+++ addons/hr_recruitment/models/__init__.py
@@ -1,5 +1,10 @@
 from . import hr_department
-from . import hr_recruitment
+from . import hr_applicant
+from . import hr_applicant_category
+from . import hr_applicant_refuse_reason
+from . import hr_recruitment_degree
+from . import hr_recruitment_source
+from . import hr_recruitment_stage
 from . import hr_employee
 from . import hr_job
 from . import res_config_settings
