+++ addons/hr_recruitment/views/hr_applicant_category_views.xml
@@ -1,0 +1,40 @@
+<?xml version="1.0"?>
+<odoo>
+    <data>
+        <record id="hr_applicant_category_view_form" model="ir.ui.view">
+            <field name="name">hr.applicant.category.form</field>
+            <field name="model">hr.applicant.category</field>
+            <field name="arch" type="xml">
+                <form string="Tags">
+                <sheet>
+                    <group>
+                        <field name="name"/>
+                        <field name="color"/>
+                    </group>
+                </sheet>
+                </form>
+            </field>
+        </record>
+
+        <record id="hr_applicant_category_view_tree" model="ir.ui.view">
+            <field name="name">hr.applicant.category.tree</field>
+            <field name="model">hr.applicant.category</field>
+            <field name="arch" type="xml">
+                <tree string="Tags" editable="bottom">
+                    <field name="name"/>
+                    <field name="color" groups="base.group_no_one"/>
+                </tree>
+            </field>
+        </record>
+
+        <record id="hr_applicant_category_action" model="ir.actions.act_window">
+            <field name="name">Tags</field>
+            <field name="res_model">hr.applicant.category</field>
+            <field name="help" type="html">
+                <p class="o_view_nocontent_smiling_face">
+                    Add a new tag
+                </p>
+            </field>
+        </record>
+    </data>
+</odoo>
