+++ addons/hr_recruitment/views/hr_applicant_refuse_reason_views.xml
@@ -1,0 +1,45 @@
+<?xml version="1.0"?>
+<odoo>
+    <data>
+        <record id="hr_applicant_refuse_reason_view_form" model="ir.ui.view">
+            <field name="name">Applicant refuse reason form</field>
+            <field name="model">hr.applicant.refuse.reason</field>
+            <field name="arch" type="xml">
+                <form string="Refuse Reason">
+                    <sheet>
+                        <widget name="web_ribbon" text="Archived" bg_color="bg-danger" attrs="{'invisible': [('active', '=', True)]}"/>
+                        <div class="oe_title">
+                            <div class="oe_edit_only">
+                                <label for="name"/>
+                            </div>
+                            <h1>
+                                <field name="name"/>
+                            </h1>
+                            <field name="active" invisible="1"/>
+                        </div>
+                        <group>
+                            <field name="template_id"/>
+                        </group>
+                    </sheet>
+                </form>
+            </field>
+        </record>
+
+        <record id="hr_applicant_refuse_reason_view_tree" model="ir.ui.view">
+            <field name="name">Applicant refuse reason tree</field>
+            <field name="model">hr.applicant.refuse.reason</field>
+            <field name="arch" type="xml">
+                <tree string="Refuse Reason" editable="bottom">
+                    <field name="name"/>
+                    <field name="template_id"/>
+                </tree>
+            </field>
+        </record>
+
+        <record id="hr_applicant_refuse_reason_action" model="ir.actions.act_window">
+            <field name="name">Refuse Reasons</field>
+            <field name="res_model">hr.applicant.refuse.reason</field>
+            <field name="view_mode">tree,form</field>
+        </record>
+    </data>
+</odoo>
