+++ addons/website_hr_recruitment/models/__init__.py
@@ -3,5 +3,6 @@
 
 from . import hr_department
 from . import hr_job
-from . import hr_recruitment
+from . import hr_applicant
+from . import hr_recruitment_source
 from . import website
